refactor: improve type safety and error handling
- Move BrochureType and LanguageStore to shared types file
- Add proper typing for BrochureForm component
- Remove unused slick-theme.css to reduce bundle size
- Enhance error handling in App component with abort controller and toasts
